Simplify path mapping of p4-fusion package
This PR is a precursor to a second PR that will drop the shell layer in between gitserver and p4-fusion.
We will track command outputs and resource requirements separately, so there's no need for this additional layer of complexity anymore.

Test plan:

This just bumps the package and simplifies the path, a follow-up PR will make use of this new epoch.