blob: Properly hide popovers when the mouse leaves the file area
There are currently two situations where the popover isn't properly
closed:

1. The popover extends outside of the file view boundaries and the mouse
  leaves the file view via the popover.
2. The hovered symbol is at the edge of the file view and the mouse
  leaves the file view moving over the symbol.

The first case is annoying becaues the popover might cover important
information and it's not possible to close the popover by moving outside
of it.

The second case is annoying because it can cause the popover to appear
when moving outside of the file view (towards the top or bottom), which
is just surprising at the best case, but can hide important information
in the worst case.

The solution implemented in this commit is to explicitly listen to
`mouseleave` events and hide the popover accordingly.