symbols: use concurrency safe map for repoToSize in rockskip (#62114)
We are getting "fatal error: concurrent map writes" in production when
enabling rockskip at large customers. The root cause is the repoToSize
map which is read for every search request (and written every "cache
miss").

This was the easiest change to make. We should probably also be single
flighting this request, but that can be changed at another time.

Additionally fixed another bug where in the cache miss case we would do
`size :=` and therefore not update size from 0, which then does a
sqliteSearch. This means on the very first request to a rockskip repo we
would always do sqliteSearch which is also a waste of resources.

Test Plan: Stefan manually testing :)