svelte: Show code intel popovers above global navigation (#62125)
On the file page it's possible that a popover is extending outside the
file view and "reaching" into the global header. Because the global
header has z-index: 1 it's currently rendered above all other, partially
covering the popover.

This commit solves this issue by removing the z-index from the global
header and instead moves it where it's needed, namely rendering search
input suggestions above search results.