Search: fix case-sensitive queries with lang filters (#62157)
This PR fixes a regression where language filters might not match when case
sensitivity is enabled. Before, we just appended `(?i)` to the file filters to
indicate case insensitivity, but still passed a case-sensitive file query to
Zoekt (called `case_file_regex`). Now, we make sure that the Zoekt queries are
also case insensitive so there is no conflict in how to interpret them.

Co-authored-by: Matthew Manela <mmanela@users.noreply.github.com>