svelte: Add minimal reference panel (#62168)
This PR adds a *very early* version of the reference panel.
Things that work:

- Fetch precise code intel information via GraphQL API
- Opening the reference panel when clicking "find references"
- "No references found" message
- Open/close file preview
- Error handling

Things that don't work yet (to be iterated on):

- Support for search-based code intel
- Grouping of references, or repo indicator in general
- Restore opened/selected preview on back/forward navigation

I want to emphasize that this is by no means a "new" reference panel. In the interest of time it's an approximation of the reference panel in the React app. Some things have been simplified, such as omitting grouping of references.

Worth noting is using the actual file page as file preview component. This is possible because pages are normal components. I'm still a bit unsure whether this is a good approach or not.

- Pro: 
  - Works nicely together with `preloadData` (except for data type casting). There is no need to duplicate the data loading logic or pluck out the right values from the preloaded data. That also means that data will be properly cached.
  - Ensures some consistency between the file page itself and wherever we show a file preview.
- Con:
  - Might make working on the page itself more difficult since the "embedded" use case has to be kept in mind.
  - Additional props are required to adjust UI for the preview case.

I'd like to give this a try and see how it works, possible also make the search results page use this for the preview panel.