Context: allow embeddings only for testing (#62162)
As we're evaluating the quality of Cody's context, it's really useful to be
able to compare against Enterprise embeddings. However, we completely disabled
Enterprise embeddings when we turned them off for customers.

This PR adds an environment variable `SRC_EMBEDDINGS_ALLOWED` that lets us
enable and test embeddings locally. It is not advertised and should never be
enabled in production use.