Context: push limiting and merging into search job (#61849)
To test and debug Cody context, it's really useful to be able to see the search
results directly. This PR makes it possible to test context in the main
Sourcegraph search bar by adding `patterntype:codycontext`.

We already have this patterntype, so the main work here is to push important
logic from the `getCodyContext` GraphQL call down into the job logic.
Specifically, it moves the logic to search code and text separately, limit the
number of files returned, and combine the results. It also moves the "Cody
ignore" filtering logic.