Context: always return code results first (#62193)
In #62162, we pushed some Cody context logic down into the 'search job'. The
tests I added assume that the results are returned in a consistent order. This
is nice behavior and makes testing and debugging easier.

I mistakenly thought that `pool.NewWithResults` would return results in the
behavior goroutines were originally submitted. Oops! Now we use two separate
wait groups and ensure we return results in a consistent order.