rockskip: skip non-utf8 paths
We seen errors on S2 for inserting non-utf8 data to the symbols table.
`path` and `name` (symbol name) are the only user controlled fields. We
already check for non-utf8 content in go-ctags, so that leaves `path`.

We simple skip the path. Which symbols within that path will simply not
exist in the index.

Test plan:
CI