rockskip: skip non-utf8 paths (#62197)
We have seen errors on S2 for inserting non-utf-8 data to the symbols table. `path` and `name` (symbol name) are the only user controlled fields. We already check for non-utf-8 content in go-ctags, so that leaves `path`, which is not required to be utf-8 compatible.

With this change we simply skip the path, which means that symbols within that path will not exist in the index. This is somewhat in line with ctags, where we skip indexing non-utf-8 content.

I believe this is an edge case and it is acceptable to skip those files to keep the code simple.

Alternatives:
- We could store arbitrary bytes `bytea` instead of `text` in Postgres.

Test plan:
CI