Context: restore Cody ignore integration test (#62220)
Follow up to #61849. We pushed Cody ignore filtering down into the Cody search
job and added tests. In the shuffle, I removed this helpful GraphQL resolvers
test, which serves as an integration test for Cody ignore config. This PR
restores it and adapts it to use a real search client to exercise the filtering
logic.