rockskip: Replace second long-running process with API
We want to migrate this call to gRPC as well. However, it isn't good practice to have to keep this process running for potentially hours, as any server restart will have to make the process start over.
With one call per commit, we're not occupying a process slot on gitserver, and rollouts of gitserver should much less affect rockskip, as individual commits can be retried before being returned to the rockskip code.

Test plan:

Existing tests still pass, I let it index a repo locally and that was fast and I get results for symbols instantly now, but not a Rockskip expert so requesting review from the owners.