Upgrade git to 2.45.0 and unpin old cURL version
We recently reverted the cURL version to 8.6.0 because it was incompatible with Git.
In the new Git release, there was a workaround put in that should make it work with latest cURL again before upstream cURL is fixed.
By that we can use the latest version again and don't need to ship a cURL with known CVEs in the next release.

Test plan:

Git integration tests are still working, E2E tests can still successfully look at repository contents, and will test that we can successfully clone the k8s repo which failed previously.