Move cody packages only used by frontend into cmd/frontend
The httpapi package for completions APIs and the internal/cody package are only used in frontend, so moving them there.

This cleans up the massive /internal package a bit, and is part of a larger effort to make less calls to conf.Get() inside /internal, because those are hard to trace down and can potentially brick services like cody-gateway when they are ever in any call paths.

Test plan:

Just moved code around, and the go compiler doesn't complain.