svelte: Add infinity scroll and search to all branches page. (#62289)
Similar to #62257. I copied most of the code because I assume we'll redesign those pages anyway, so there is no point spending time to make these components reusable.

I implemented it such that the content is centered and the scrollbar is at the right edge of the screen. I added some padding so that there is some space between edge and content on smaller screens (I made the same change to the tags page).

I also updated the `GitReference` component to use `Badge` instead, which also affects the tags page.