svelte: Update some aspects diff view and history panel (#62316)
"View at commit" on a directory page didn't work at all. I first added
support for it but it become clear that it's not obvious how such a view
would behave when clicking on a directory entry. So I decided to remove
it from this view, for now.

I also moved the "Close x" action to the info header (after talking to
Taiyab).