rockskip: Switch from long-running process to pagination (#62259)
We want to migrate this call to gRPC. However, it isn't good practice to have to keep this process running for potentially hours, as any server restart will have to make the process start over.
With pagination here, we're not occupying a process slot on gitserver, and rollouts of gitserver should much less affect rockskip, as individual pages can be retried before being returned to the rockskip code.

Test plan:

Existing tests still pass, but not a Rockskip expert so requesting review from the owners.