gitserver: Implement CommitGraph using Commits
This PR consolidates the client methods CommitGraph and Commits into one, to reduce the surface of APIs we have to migrate to gRPC and then maintain.
To my best understanding, this should be equivalent, just use a different API.

Test plan:

Existing test suites and integration tests are still passing.