gitserver: Utilize Commits method for RevList
This PR replaces the gitserver client method RevList by the Commits method.
The Commits method can do the same, but it saves us from implementing and testing a second API. Also, the name RevList sounded very generic but it was super opinionated in the flags it uses, so it was really "ListCommitsForRockskip" but that is too application-specific so Commits is probably just as good.
We'll learn a bunch about our API endpoints as we finish migrating them all, so we can do proper optimizations afterwards.

Closes https://github.com/sourcegraph/sourcegraph/issues/62098

Test plan:

Integration and unit test for rockskip are still passing.