svelte: Improve tooltip target logic (#62373)
When I encountered #62372 I also noticed that the tooltip on the smaller
avatar component is not positioned correctly. That's because the first
element child does not contribute to the layout when the style prop is
set.
I don't know how often we'll run into this situation but it seemed
reasonable to make the component a bit more resilient for such cases.

* Remove tooltip test

I assume the test fails because it's not actually rendering the elements.