svelte: Improve file sidebar's scroll into view behavior
At the moment the selected entry is scrolled to the top or bottom of the
sidebar, where it is not always that visible.
This commit changes the behavior so that it it's scrolled to the center
when the selected entry changes 'indirectly' (e.g. going back and forth
via browser history buttons).