chore: Group upload matching options into struct (#62335)
There are lots of layers of indirection/plumbing for propagating
the various options/settings for matching uploads. There was also
a mix of terminology - 'toolName' in some places and 'indexer'
in some other places. 'exactPath' in some places and 'rootMustEnclosePath'
in other places. Grouping the various options into a struct provides
a central place to document how the path matching happens,
as well as special arguments such as `shared.SyntacticIndexer`.

This also makes it much easier to understand where the various
upload matching options are used via 'Find usages' in the editor.

As a nice side-effect, the noise of creating o11y-related K-V pairs
is also reduced.