fix: Avoid locating incorrect uploads during filtering
The upload filtering code previously did a check whether:
- The commit corresponding to the upload actually existing on gitserver
- The path of interest had an associated SCIP document in that upload.

However, the second code path incorrectly only handled the case when
the first operation was a no-op. In case some uploads were filtered
out due to the path existing (or not) in the wrong upload, and the
final result could still have uploads for which the commit no longer
existed on gitserver.