svelte: Improve page title updates
I originally thought we have to create the page title in data loaders
and set it only in the root layout, but it appears that setting the
title in the layout files is what's potentially problematic (sometimes
the page title "reverts" to what's being set by the root layout when
navigating to a new page).

Only setting the page title in `+page.svelte` files seems to solve the
issue. At least for now I prefer that because it makes setting the page
title very explicit.

I also changed the file preview logic to use the (newly refactored)
component instead of the page component. Otherwise opening the preview
updates the page title, which is not correct.

 Note: One might think that having the title set in loaders would allow
 us to "compose" the title hierarchically more easily. In theory that's
 true, but having to `await` the parent's data is annoying too.