svelte: Give last commit message as much space as possible
I don't see a reason why we need to restrict the width to a specific
pixel value, especially if there would be enough space otherwise. There
are already conventions about the size of the subject line so most of
the time the message won't take up a lot of space anyway.

We could still restrict it to a relative value of the available width if
we wanted to.