Executors: run script as script, not as command (#62466)
This fixes an issue that is generating failures on repos with spaces in their names.

Previously, we were passing the script name as a shell command with -c. However, that means it's subject to shell escaping. If we instead pass the file as a script for sh to run, it avoids the need to escape the file name and also reduces the surface area for injection attacks.