appliance: container image config (#62442)
* appliance: refactor container config

In preparation for adding a field for image overrides, move the
resources section into a new ContainerConfig field, to avoid cluttering
the top-level interface.

* appliance: container image config

Add image defaults and overrides similar to the helm chart: we hardcode
our standard image repository and version-specific tags for each
component as defaults. We expose per-service overrides of the image and
its tag (although _not_ qualified by repository), and a global override
for the repository.

This commit does not support per-service overrides of the repository,
although this is something we could add in the future if there is
demand.

The appliance will need to support a range of SG versions. This commit
adds the beginnings of the scaffolding for that. Most k8s resource
manifest content for most services will not change at all betwen
subequent releases, but all image tags will likely change with every
release. Therefore, we centralize the image management across services
in one file, rather than leaving the image:version mapping to each
service. Service configs can opt out of this mechanism by not calling
getDefaultImage().