gitserver: Always use Commits in HasCommitAfter
There used to be two implementations of this, one for with subrepo enabled, one for without. This PR aligns those two to both use the Commits method.
One code path less, and one API less that we need to port to gRPC.

Test plan:

Existing unit tests are still passing.