svelte: Reduce imported global CSS (#62457)
Until now we've simply imported the global `base.scss` file, which in
turn imports a bunch of other files and defines global CSS classes. Many
(most?) of them are actually not used.

This commit reduces to the number of imported files to the ones that
contain styles that are used. My methodology was to inspect every
file that `base.scss` imports and find out whether it defines custom
properties (variables) or global classes. In some cases I extracted
variables into their own files.