gitserver: Experiment with moving commit search out of gitserver
The search is CPU and memory/allocation intensive, so we want to experiment
with moving the code path to searcher which can scale horizontally with low
latency vs gitservers very static sharding.

This needs more work on the APIs front, as the existing APIs don't 100% satisfy
the use-case here.

Test plan:

TBD