gitserver: Add API to enumerate repos on disk
This can be useful for debugging as it can give a hint on the inferred name from the path on disk, and will come in useful when we start on coordinator.
I want to use this to experiment a bit with dotcom, and the change seems fairly uninvasive, if it turns out to not be useful, I can remove it again later.

Test plan:

Added unit tests to verify pagination is working correctly.