gitserver: Implement RefHash in backend
This PR moves the implementation of ComputeRefHash to the GitBackend interface and reduces the amount of memory needed by it. We get proper tracking of the command by that, including error logs, resource usage, invocation count etc.

Since there is no 1000% guarantee that this didn't change the hash, I opted for resetting the last changed value and recomputing from scratch, by changing the path of the file to _v2.

Test plan:

Tests are still passing.