gitserver: Reduce number of git config invocations
Using config is fairly convenient, but got somewhat out of hand. On dotcom, we call this millions of times, because the janitor runs frequently and there are a lot of repos.
Instead, for a few simple things we now use a different mechanism that has less overhead (also, we only abused the git config, those values are of no interest to git itself).
1. The timestamp for the last size calculation is now stored as the mtime of a magic file in the repo dir.
2. The repoType is only loaded when there is a reclone reason, and we might want to skip over Perforce repos

Test plan:

Added a basic test, and integration tests still pass.