gitserver: Fix duration reported for ChangedFiles (#62603)
* gitserver: Make ReadDir return an iterator on client side

There is potentially a lot of FDs in a repo, so we return an iterator instead to encourage callers to consume it one by one and not load all of it into memory.

Test plan:

Existing tests still pass.

* gitserver: Fix duration reproted for ChangedFiles

We always ended the observation when the iterator is created before, this will make sure to track the time until the call to Close().

Test plan:

Code review, existing CI.