gitserver: Move repo timestamp updates to caller (#62610)
This is a small step towards making gitserver itself less dependent on Sourcegraph concepts like the database, by making the repo-updater scheduler (later, this can be coordinator) write to the DB instead.

Note: Until we decide what to do with ensurerevision which schedules out-of-band updates that affect concurrency and the control flow, I decided to inline this behavior there to make it very explicit.

Test plan:

Existing tests are still passing.