Python notebooks: fix anchor offset (#62637)
This fixes an issue with Jupyter notebook rendering where the offset from the anchor can clip the left side of the header text. The issue is that we use margin-left: 1rem; to offset the anchor, but the anchor isn't guaranteed to be exactly 1rem wide. This makes the relationship explicit and adds a comment that those are linked.