gitserver: Allow multiple ranges to be passed to Commits
This will be useful for commit searches, and is generally supported by git so why not support it here. That also lets you express `A..B` as `B, ^A` if you want to do less string concatenation, for example.

Test plan:

Made sure all existing callers pass a non-empty string to the slice if they specify the slice at all, and all existing tests still pass.