observability: Fix error checks in ErrorsFilter
The ErrorsFilter function can be used to skip over well known errors for certain log targets. Multiple errors can all be collected in a wrapper, the ErrCollector.
The collector wraps the errors it found, but it doesn't implement Unwrap, so errors.Is and errors.HasType don't work properly here, they will only see the ErrCollector.
This PR adds Unwrap to the ErrCollector to fix that, and fixes the error code detection in gitserver git observability, and makes it less aggressive in what we skip.

Test plan:

Verified that the two errors are no longer erroneously logged.