svelte: Show correct keyboard shortcut on Linux in "search files" button
This commit updates the repo sidebar to show the Linux specific keyboard
shorcut for searching files.

I initially was just rendering it as a string, but I think makes sense
to have some spacing between the individual parts, just like it was done
in the sidebar. So I added a new component for rendering
platform-specific keyboard shortcuts that applies this style.

The current code adds some negative padding, I don't know if that's
really necessary.