gitserver: Fix Diff generation
When switching to git-diff-tree, I looked over the fact that it can only work with treeish-es, not with rangespecs.

This PR translates ranges to the corresponding equivalents according to the Git docs:

`..` -> `A B`
`...` -> `$(git merge-base A B) B`

Test plan:

Verified diffs are now generated properly again, also added another test to our suite.