authz: Fix panic when auth provider is missing
When authz is enabled for a GitHub code host connection, but there is no corresponding auth.providers entry, this code currently panics, because `GetOAuthContext` returns nil, and then the `RefreshFunc` in the OAuthBearerToken tries to use that context when it refreshes.

This is a misconfiguration, and cannot work (we need both the authz setting, and the auth provider), but the worker pod should not just panic because of that, so we return a nicer error here now.

Test plan:

Added a unit test.