repoupdater: Add retry policy to repositoryservice
This new client didn't have any retries set, so we would always fast-fail when gitserver is unavailable. That makes the operation more noisy than it has to be, and potentially all repos back in the fetch queue, because it thinks the fetch itself failed.
All these methods are safe to be called a second time, so we can safely do that.

This is an easy fix to ensure repo updater waits a bit for gitserver to be back online.

Test plan:

Verified locally that fetches don't immediately fail when gitserver is down.