svelte: Improve commit page loading state
It was reported that we show the old commit data when navigating to a
different commit page.
The page was explicitly implemented to show the previously loaded data
in case there is an error when fetching more data. But this also affects
normal page navigation.

In this commit the implementation was changed to return the previously
loaded data when an error occurs, which simplifies loading state and
data rendering on the page itself.